+++ src/options/presets/MediumObfuscation.ts
@@ -16,6 +16,7 @@ export const MEDIUM_OBFUSCATION_PRESET: TInputOptions = Object.freeze({
     stringArrayEncoding: [
         StringArrayEncoding.Base64
     ],
-    stringArrayIntermediateVariablesCount: 5,
+    stringArrayWrappersChainedCalls: true,
+    stringArrayWrappersCount: 2,
     transformObjectKeys: true
 });